Split TimeOnlyAssertionSpecs
